Changes from blimp to robot
Changed pins from D9, D10 to 18,19 to clean up wire routing so it goes on one side of the board.

Removed lift stepper motor and replaced with blade motor (2).  Mapped forward and steering to motors 3 and 4.

I'm realizing this doesn't have a reverse, so I'll need to remap values, and steering might still need some work.